refactor(runtime): split accessors.rs into cohesive modules (§7-8)
`accessors.rs` was 2298 lines, well over the 500-line module limit. Split it
into six files along its functional seams, each under 500 lines, with no
behavior change (the set of fn signatures is identical before/after, modulo
visibility):

- `accessors.rs` (487) — return-type & coercion: callable return-type/spec
  resolution, definite-return slot handling, Failure↔error conversion, return
  type-check/coercion enforcement, and `where`-constraint checks.
- `accessors_state.rs` (489) — callable signature/composition and interpreter
  state accessors: `our`/`state`/once vars, closure-captured state, wrap chains,
  and method/multi/proto/samewith dispatch-frame stacks.
- `accessors_resolve.rs` (372) — method-body compilation
  (`compile_method_def_in_place`/`compile_class_methods`), `resolve_code_var`,
  and smart-match/sequence eval.
- `accessors_stack.rs` (192) — routine/block/gather execution-stack accessors
  and current-package state.
- `accessors_stash.rs` (486) — symbolic stash member lookup and
  package/indirect-type-name resolution (`resolve_indirect_type_name`,
  `package_stash_value`).
- `accessors_misc.rs` (303) — misc interpreter state: my-global-stash
  visibility, `is_pseudo_package_name`, END phasers, routine-registry
  snapshot/restore, block-scope depth, and `let`-saves.

`normalize_categorical_operator_name` and `inferred_operator_arity` are bumped
`fn`→`pub(super)` (called from the resolve/state modules); `should_hide_from_my_global_stash`
and `is_pseudo_package_name` (both `pub(crate)`) move to `accessors_misc`. Each
file carries only the imports it uses (`accessors_stash`/`accessors_stack` drop
the unused `Symbol`). Eleventh slice of the §7-8 file-splitting sweep.
