+++ src/runtime/accessors.rs
@@ -2,34 +2,13 @@ use super::*;
 use crate::symbol::Symbol;
 
 impl Interpreter {
-    pub(crate) fn should_hide_from_my_global_stash(&self, key: &str) -> bool {
-        if key.starts_with('$')
-            || key.starts_with('@')
-            || key.starts_with('%')
-            || key.starts_with('&')
-        {
-            return false;
-        }
-        self.need_hidden_classes.contains(key)
-            || key
-                .strip_prefix("GLOBAL::")
-                .is_some_and(|name| self.need_hidden_classes.contains(name))
-            || key
-                .rsplit_once("::")
-                .is_some_and(|(_, short)| self.need_hidden_classes.contains(short))
-            || self
-                .need_hidden_classes
-                .iter()
-                .any(|name| key.ends_with(&format!("::{name}")))
-    }
-
-    fn normalize_categorical_operator_name(name: &str) -> String {
+    pub(super) fn normalize_categorical_operator_name(name: &str) -> String {
         // Keep categorical operator names as-written. Parenthesized operators
         // like infix:<(==)> are distinct from infix:<==>.
         name.to_string()
     }
 
-    fn inferred_operator_arity(name: &str) -> Option<usize> {
+    pub(super) fn inferred_operator_arity(name: &str) -> Option<usize> {
         if name.starts_with("infix:<") && name.ends_with('>') {
             return Some(2);
         }
@@ -503,1796 +482,5 @@ impl Interpreter {
             Ok(result) => result.truthy(),
             Err(_) => false,
         }
-    }
-
-    pub(crate) fn callable_signature(&self, callable: &Value) -> (Vec<String>, Vec<ParamDef>) {
-        match callable {
-            Value::Sub(data) => (data.params.clone(), data.param_defs.clone()),
-            Value::Routine { name, .. } => {
-                if let Some(def) = self.resolve_function(&name.resolve()) {
-                    return (def.params.clone(), def.param_defs.clone());
-                }
-                if let Some(arity) = Self::inferred_operator_arity(&name.resolve()) {
-                    let params = (0..arity).map(|i| format!("arg{}", i)).collect();
-                    return (params, Vec::new());
-                }
-                // Well-known 0-arity terms, and slurpy (`*@args`) builtins whose
-                // required-positional arity is 0, should report no parameters
-                // (raku: `&warn.arity == 0`). Mustache's logger keys on
-                // `&warn.?arity == 2`, so a spurious arity of 1 sent it down the
-                // wrong (2-arg) branch.
-                if matches!(
-                    name.resolve().as_ref(),
-                    "rand"
-                        | "now"
-                        | "time"
-                        | "warn"
-                        | "note"
-                        | "say"
-                        | "print"
-                        | "put"
-                        | "die"
-                        | "fail"
-                ) {
-                    return (Vec::new(), Vec::new());
-                }
-                (vec!["arg0".to_string()], Vec::new())
-            }
-            _ => (vec!["arg0".to_string()], Vec::new()),
-        }
-    }
-
-    pub(crate) fn infix_associativity(&self, full_name: &str) -> Option<String> {
-        let fq = format!("{}::{}", self.current_package(), full_name);
-        self.operator_assoc
-            .get(&fq)
-            .cloned()
-            .or_else(|| self.operator_assoc.get(full_name).cloned())
-            .or_else(|| {
-                let global = format!("GLOBAL::{}", full_name);
-                self.operator_assoc.get(&global).cloned()
-            })
-    }
-
-    pub(crate) fn call_user_routine_direct(
-        &mut self,
-        full_name: &str,
-        args: Vec<Value>,
-    ) -> Result<Value, RuntimeError> {
-        if let Some(def) = self.resolve_function_with_alias(full_name, &args) {
-            return self.call_function_def(&def, &args);
-        }
-        if let Some(err) = self.take_pending_dispatch_error() {
-            return Err(err);
-        }
-        let env_name = format!("&{}", full_name);
-        if let Some(callable) = self.env.get(&env_name).cloned() {
-            return self.eval_call_on_value(callable, args);
-        }
-        // Try stripping package prefix (e.g., "Main::foo" -> "foo")
-        // when the function is in the current or GLOBAL package.
-        if let Some(pos) = full_name.rfind("::") {
-            let short_name = &full_name[pos + 2..];
-            if let Some(def) = self.resolve_function_with_alias(short_name, &args) {
-                return self.call_function_def(&def, &args);
-            }
-            let env_short = format!("&{}", short_name);
-            if let Some(callable) = self.env.get(&env_short).cloned() {
-                return self.eval_call_on_value(callable, args);
-            }
-        }
-        let suggestions = self.suggest_routine_names(full_name);
-        Err(RuntimeError::undeclared_routine_symbols(
-            full_name,
-            format!("Unknown function: {}", full_name),
-            suggestions,
-        ))
-    }
-
-    pub(crate) fn compose_callables(&self, left: Value, right: Value) -> Value {
-        use std::sync::atomic::{AtomicU64, Ordering};
-
-        static COMPOSE_ID: AtomicU64 = AtomicU64::new(1_000_000);
-        let id = COMPOSE_ID.fetch_add(1, Ordering::Relaxed);
-
-        let (mut params, param_defs) = self.callable_signature(&right);
-        if params.is_empty() {
-            if !param_defs.is_empty() {
-                params = param_defs.iter().map(|pd| pd.name.clone()).collect();
-            } else {
-                params = vec!["arg0".to_string()];
-            }
-        }
-
-        let left_return_type = self.callable_return_type(&left);
-        let mut env = crate::env::Env::new();
-        env.insert("__mutsu_compose_left".to_string(), left);
-        env.insert("__mutsu_compose_right".to_string(), right);
-        if let Some(rt) = left_return_type {
-            env.insert("__mutsu_return_type".to_string(), Value::str(rt));
-        }
-
-        Value::make_sub_with_id(
-            Symbol::intern(""),
-            Symbol::intern("<composed>"),
-            params,
-            param_defs,
-            Vec::new(),
-            false,
-            env,
-            id,
-        )
-    }
-
-    pub(crate) fn env_mut(&mut self) -> &mut Env {
-        &mut self.env
-    }
-
-    /// Get a cloned copy of the persisted closure env for a given closure id.
-    pub(crate) fn get_closure_env_override(&self, id: u64) -> Option<crate::env::Env> {
-        self.closure_env_overrides.get(&id).cloned()
-    }
-
-    /// Check whether a sub has an active (non-empty) wrap chain.
-    pub(crate) fn has_wrap_chain(&self, sub_id: u64) -> bool {
-        self.wrap_chains.get(&sub_id).is_some_and(|c| !c.is_empty())
-    }
-
-    /// Check whether we're already inside a wrap dispatch for a given sub.
-    pub(crate) fn is_wrap_dispatching(&self, sub_id: u64) -> bool {
-        self.wrap_dispatch_stack.iter().any(|f| f.sub_id == sub_id)
-    }
-
-    /// Find the sub_id and Sub value for a function name that has an active wrap chain.
-    /// Returns the sub_id if a wrap chain exists for the given function name.
-    pub(crate) fn wrap_sub_id_for_name(&self, name: &str) -> Option<u64> {
-        for (sub_id, sub_name) in &self.wrap_sub_names {
-            if sub_name == name && self.has_wrap_chain(*sub_id) {
-                return Some(*sub_id);
-            }
-        }
-        None
-    }
-
-    /// Get the original wrapped Sub value for a function name.
-    /// Returns the Sub value stored when wrap was called, preserving the original sub_id.
-    pub(crate) fn get_wrapped_sub(&self, name: &str) -> Option<Value> {
-        self.wrap_name_to_sub.get(name).cloned()
-    }
-
-    pub(crate) fn get_our_var(&self, key: &str) -> Option<&Value> {
-        self.our_vars.get(key)
-    }
-
-    pub(crate) fn our_vars_iter(&self) -> impl Iterator<Item = (&String, &Value)> {
-        self.our_vars.iter()
-    }
-
-    pub(crate) fn set_our_var(&mut self, key: String, value: Value) {
-        self.our_vars.insert(key, value);
-    }
-
-    pub(crate) fn get_state_var(&self, key: &str) -> Option<&Value> {
-        self.state_vars.get(key)
-    }
-
-    pub(crate) fn set_state_var(&mut self, key: String, value: Value) {
-        self.state_vars.insert(key, value);
-    }
-
-    /// Track C: get-or-create a shared `ContainerRef` cell for a `state` variable
-    /// in `shared_vars` (the cross-thread store), keyed by `key`. Used while a
-    /// thread is running so that concurrent calls to the same routine — e.g.
-    /// `await (^3).map: { start f() }` where `f` has `state $n` — share one live
-    /// cell instead of each thread initializing its own snapshot. The get-or-init
-    /// is atomic under the `shared_vars` write lock, so the first caller seeds the
-    /// cell (from `initial`) and the rest observe it. Returns the cell value.
-    pub(crate) fn get_or_init_shared_state_cell(&self, key: &str, initial: Value) -> Value {
-        let mut sv = self.shared_vars.write().unwrap();
-        if let Some(existing) = sv.get(key)
-            && matches!(existing, Value::ContainerRef(_))
-        {
-            return existing.clone();
-        }
-        let cell = initial.into_container_ref();
-        sv.insert(key.to_string(), cell.clone());
-        cell
-    }
-
-    /// Read per-closure-instance captured-variable state (hot closure-call path).
-    pub(crate) fn get_closure_captured_state(&self, id: u64, name: Symbol) -> Option<&Value> {
-        self.closure_captured_state.get(&(id, name))
-    }
-
-    /// Persist per-closure-instance captured-variable state (hot closure-call path).
-    pub(crate) fn set_closure_captured_state(&mut self, id: u64, name: Symbol, value: Value) {
-        self.closure_captured_state.insert((id, name), value);
-    }
-
-    pub(crate) fn current_once_scope(&self) -> Option<u64> {
-        // When inside a closure call, __mutsu_callable_id identifies the
-        // specific closure clone.  This must take priority over the
-        // once_scope_stack so that `once` blocks inside closures called from
-        // EVAL (where the stack depth > 1) still get a per-clone scope.
-        match self.env.get("__mutsu_callable_id") {
-            Some(Value::Int(id)) if *id >= 0 => Some(*id as u64),
-            _ => self.once_scope_stack.last().copied(),
-        }
-    }
-
-    pub(crate) fn push_once_scope(&mut self, scope: u64) {
-        self.once_scope_stack.push(scope);
-    }
-
-    pub(crate) fn pop_once_scope(&mut self) {
-        self.once_scope_stack.pop();
-    }
-
-    pub(crate) fn next_once_scope_id(&mut self) -> u64 {
-        let scope = self.next_once_scope_id;
-        self.next_once_scope_id += 1;
-        scope
-    }
-
-    pub(crate) fn get_once_value(&self, key: &str) -> Option<&Value> {
-        self.once_values.get(key)
-    }
-
-    pub(crate) fn set_once_value(&mut self, key: String, value: Value) {
-        self.once_values.insert(key, value);
-    }
-
-    pub(crate) fn when_matched(&self) -> bool {
-        self.when_matched
-    }
-
-    pub(crate) fn set_when_matched(&mut self, v: bool) {
-        self.when_matched = v;
-    }
-
-    pub(crate) fn is_role(&self, name: &str) -> bool {
-        self.registry().roles.contains_key(name)
-    }
-
-    pub(crate) fn push_method_class(&mut self, class_name: String) {
-        self.method_class_stack.push(class_name);
-    }
-
-    pub(crate) fn pop_method_class(&mut self) {
-        self.method_class_stack.pop();
-    }
-
-    pub(crate) fn method_class_stack_top(&self) -> Option<String> {
-        self.method_class_stack.last().cloned()
-    }
-
-    /// Set up a method dispatch frame for nextsame/callsame support.
-    /// Returns true if a frame was pushed (caller must call pop_method_dispatch).
-    /// Also pushes a samewith context unconditionally for samewith() support.
-    pub(crate) fn push_method_dispatch_frame(
-        &mut self,
-        receiver_class: &str,
-        method_name: &str,
-        args: &[Value],
-        invocant: Value,
-    ) -> bool {
-        // Always push samewith context so samewith() can find the method name/invocant
-        self.samewith_context_stack
-            .push((method_name.to_string(), Some(invocant.clone())));
-        let all_candidates = self.resolve_all_methods_with_owner(receiver_class, method_name, args);
-        // Identify the chosen candidate and skip exactly that one
-        let chosen = self.resolve_method_with_owner(receiver_class, method_name, args);
-        let chosen_fp = chosen.as_ref().map(|(_, def)| {
-            crate::ast::function_body_fingerprint(&def.params, &def.param_defs, &def.body)
-        });
-        let mut remaining: Vec<(String, super::MethodDef)> = Vec::new();
-        let mut skipped_chosen = false;
-        for (owner, def) in all_candidates {
-            let fp = crate::ast::function_body_fingerprint(&def.params, &def.param_defs, &def.body);
-            if !skipped_chosen && Some(fp) == chosen_fp {
-                skipped_chosen = true;
-                continue;
-            }
-            if self.should_skip_defer_method_candidate(receiver_class, &owner) {
-                continue;
-            }
-            remaining.push((owner, def));
-        }
-        let pushed = !remaining.is_empty();
-        if pushed {
-            let rw_params = chosen
-                .as_ref()
-                .map(|(_, def)| {
-                    super::builtins_dispatch_next::rw_scalar_positional_params(&def.param_defs)
-                })
-                .unwrap_or_default();
-            self.method_dispatch_stack.push(super::MethodDispatchFrame {
-                receiver_class: receiver_class.to_string(),
-                invocant,
-                args: args.to_vec(),
-                remaining,
-                rw_params,
-            });
-        }
-        pushed
-    }
-
-    pub(crate) fn is_inside_wrap_dispatch(&self) -> bool {
-        !self.wrap_dispatch_stack.is_empty()
-    }
-
-    pub(crate) fn has_any_wrap_chains(&self) -> bool {
-        !self.method_wrap_chains.is_empty()
-    }
-
-    pub(crate) fn push_wrap_dispatch_frame(&mut self, frame: super::WrapDispatchFrame) {
-        self.wrap_dispatch_stack.push(frame);
-    }
-
-    pub(crate) fn pop_wrap_dispatch_frame(&mut self) {
-        self.wrap_dispatch_stack.pop();
-    }
-
-    /// Get method-level wrap chain for a specific candidate.
-    pub(crate) fn get_method_wrap_chain(
-        &self,
-        class_name: &str,
-        method_name: &str,
-        candidate_idx: usize,
-    ) -> Option<&Vec<(u64, Value)>> {
-        let key = (
-            class_name.to_string(),
-            method_name.to_string(),
-            candidate_idx,
-        );
-        self.method_wrap_chains.get(&key).filter(|c| !c.is_empty())
-    }
-
-    /// Find the candidate index for a method definition in its class.
-    pub(crate) fn find_method_candidate_index(
-        &self,
-        class_name: &str,
-        method_name: &str,
-        method_def: &super::MethodDef,
-    ) -> Option<usize> {
-        // No user-code re-entry here, so a let-bound guard is safe.
-        let registry = self.registry();
-        let class_def = registry.classes.get(class_name)?;
-        let defs = class_def.methods.get(method_name)?;
-        let target_fp = crate::ast::function_body_fingerprint(
-            &method_def.params,
-            &method_def.param_defs,
-            &method_def.body,
-        );
-        defs.iter().position(|d| {
-            crate::ast::function_body_fingerprint(&d.params, &d.param_defs, &d.body) == target_fp
-        })
-    }
-
-    /// Pop a method dispatch frame (must only be called if push returned true).
-    pub(crate) fn pop_method_dispatch(&mut self) {
-        self.method_dispatch_stack.pop();
-    }
-
-    /// Pop the samewith context pushed by push_method_dispatch_frame.
-    /// Must always be called after push_method_dispatch_frame, regardless of its return value.
-    pub(crate) fn pop_method_samewith_context(&mut self) {
-        self.samewith_context_stack.pop();
-    }
-
-    /// Push a multi dispatch frame for callsame/nextsame/callwith/nextwith support.
-    /// Returns true if a frame was pushed (i.e. there are remaining candidates).
-    pub(crate) fn push_multi_dispatch_frame(&mut self, name: &str, args: &[Value]) -> bool {
-        // Collect ALL multi candidates regardless of arg matching. This is
-        // needed because callwith() can re-dispatch with different args, so
-        // candidates that don't match the original args may match the new ones.
-        let all_candidates = self.resolve_all_multi_candidates(name);
-        if all_candidates.len() <= 1 {
-            return false;
-        }
-        // Identify the candidate currently being called by the DETERMINISTIC
-        // dispatch winner (the same resolver the interpreter's inline frame uses),
-        // NOT a HashMap-ordered first match. `resolve_all_matching_candidates` is
-        // HashMap-ordered, so its `.first()` is not reliably the winner: when the
-        // narrowest candidate is declared after a broader one, callsame/nextsame
-        // would redispatch to the wrong (or the same) candidate, flaking ~50% of
-        // the time with the process hash seed. The winner is excluded from
-        // `remaining` so redispatch targets the OTHER candidates.
-        let saved_err = self.take_pending_dispatch_error();
-        let current_fp = self.resolve_function_with_alias(name, args).map(|def| {
-            crate::ast::function_body_fingerprint(&def.params, &def.param_defs, &def.body)
-        });
-        if let Some(err) = saved_err {
-            self.set_pending_dispatch_error(err);
-        }
-        let remaining: Vec<std::sync::Arc<super::FunctionDef>> = all_candidates
-            .into_iter()
-            .filter(|c| {
-                let fp = crate::ast::function_body_fingerprint(&c.params, &c.param_defs, &c.body);
-                Some(fp) != current_fp
-            })
-            .collect();
-        let pushed = !remaining.is_empty();
-        if pushed {
-            // Capture the FIRST (winning) candidate's scalar rw params so a
-            // nextsame+rw redispatch can chain the rw value through it (§D).
-            let rw_params = self
-                .resolve_function_with_alias(name, args)
-                .map(|def| {
-                    super::builtins_dispatch_next::rw_scalar_positional_params(&def.param_defs)
-                })
-                .unwrap_or_default();
-            self.multi_dispatch_stack
-                .push((name.to_string(), remaining, args.to_vec(), rw_params));
-        }
-        pushed
-    }
-
-    /// Pop a multi dispatch frame (must only be called if push returned true).
-    pub(crate) fn pop_multi_dispatch(&mut self) {
-        self.multi_dispatch_stack.pop();
-    }
-
-    /// Push a proto-sub dispatch frame so a compiled proto body's `{*}`
-    /// (`__PROTO_DISPATCH__`) can read the original proto args when it
-    /// redispatches to the winning multi candidate (ledger §D).
-    pub(crate) fn push_proto_dispatch_frame(&mut self, name: String, args: Vec<Value>) {
-        self.proto_dispatch_stack.push((name, args, None));
-    }
-
-    /// Pop the proto-sub dispatch frame pushed by `push_proto_dispatch_frame`.
-    pub(crate) fn pop_proto_dispatch_frame(&mut self) {
-        self.proto_dispatch_stack.pop();
-    }
-
-    /// Clone of the current proto-dispatch frame `(name, args, method_ctx)`, read
-    /// by the VM-native `{*}` redispatch handler. `None` outside a proto body.
-    #[allow(clippy::type_complexity)]
-    pub(crate) fn proto_dispatch_last(
-        &self,
-    ) -> Option<(String, Vec<Value>, Option<super::ProtoMethodCtx>)> {
-        self.proto_dispatch_stack.last().cloned()
-    }
-
-    /// Push a samewith context for a multi sub dispatch.
-    pub(crate) fn push_samewith_context(&mut self, name: &str, invocant: Option<Value>) {
-        self.samewith_context_stack
-            .push((name.to_string(), invocant));
-    }
-
-    /// Pop a samewith context.
-    pub(crate) fn pop_samewith_context(&mut self) {
-        self.samewith_context_stack.pop();
-    }
-
-    #[allow(dead_code)]
-    pub(crate) fn class_composed_roles(&self, class_name: &str) -> Option<Vec<String>> {
-        self.registry()
-            .class_composed_roles
-            .get(class_name)
-            .cloned()
-    }
-
-    #[allow(dead_code)]
-    pub(crate) fn get_role_def(&self, role_name: &str) -> Option<super::RoleDef> {
-        self.registry().roles.get(role_name).cloned()
-    }
-
-    pub(crate) fn class_role_param_bindings(
-        &self,
-        class_name: &str,
-    ) -> Option<HashMap<String, Value>> {
-        self.registry()
-            .class_role_param_bindings
-            .get(class_name)
-            .cloned()
-    }
-
-    /// Compile all uncompiled method bodies in a method map.
-    /// Compile a single method/submethod body to bytecode in place if it is not
-    /// already compiled. Shared by the bulk registration pass and the on-demand
-    /// compile in `run_resolved_method_compiled_or_treewalk`. An empty body is
-    /// compiled too (to a trivial body returning Nil/self) so that empty `BUILD`/
-    /// `TWEAK`/`method foo {}` stubs no longer fall through to the tree-walk
-    /// `run_instance_method_resolved` — leaving only delegation forwarders there.
-    pub(crate) fn compile_method_def_in_place(def: &mut super::MethodDef, package_name: &str) {
-        // A delegation forwarder (`handles`) has a synthesized/empty body and must
-        // keep its delegation routing — compiling its empty body would make paths
-        // that check `compiled_code.is_some()` run the empty body (returning Nil)
-        // instead of forwarding.
-        if def.compiled_code.is_some() || def.delegation.is_some() {
-            return;
-        }
-        let mut compiler = crate::compiler::Compiler::new();
-        let method_package = def
-            .original_role
-            .as_deref()
-            .or(def.role_origin.as_deref())
-            .unwrap_or(package_name);
-        compiler.set_current_package(method_package.to_string());
-        // A method always carries an implicit `*%_` / `*@_` slurpy, so `%_` / `@_`
-        // are valid lexicals throughout the body (including a nested signature-less
-        // `do {}` block). Mark the method context so the do-block placeholder check
-        // permits them.
-        compiler.lexically_in_method = true;
-        let mut method_params = vec![
-            "self".to_string(),
-            "__ANON_STATE__".to_string(),
-            "?CLASS".to_string(),
-            "?ROLE".to_string(),
-        ];
-        method_params.extend(def.params.iter().cloned());
-        let mut cc =
-            compiler.compile_routine_closure_body(&method_params, &def.param_defs, &def.body);
-        cc.compute_may_capture_outer_vars();
-        cc.compute_needs_env_sync();
-        def.compiled_code = Some(std::sync::Arc::new(cc));
-    }
-
-    fn compile_methods_for_map(
-        methods: &mut HashMap<String, Vec<super::MethodDef>>,
-        package_name: &str,
-    ) {
-        for overloads in methods.values_mut() {
-            for def in overloads.iter_mut() {
-                Self::compile_method_def_in_place(def, package_name);
-            }
-        }
-    }
-
-    /// Check all methods in a class for assignment to native-typed read-only
-    /// parameters. Returns the first error found, or None.
-    pub(crate) fn check_class_native_readonly_param_errors(
-        &self,
-        class_name: &str,
-    ) -> Option<crate::value::RuntimeError> {
-        // No user-code re-entry here (only the static compiler check runs), so a
-        // let-bound guard is safe.
-        let registry = self.registry();
-        let class_def = registry.classes.get(class_name)?;
-        for overloads in class_def.methods.values() {
-            for def in overloads {
-                if let Some(err_val) =
-                    crate::compiler::Compiler::check_native_readonly_param_assignment(
-                        &def.param_defs,
-                        &def.body,
-                    )
-                {
-                    let msg = if let Value::Instance { attributes, .. } = &err_val {
-                        attributes
-                            .as_map()
-                            .get("message")
-                            .map(|v| v.to_string_value())
-                            .unwrap_or_else(|| "Cannot assign to readonly variable".to_string())
-                    } else {
-                        "Cannot assign to readonly variable".to_string()
-                    };
-                    let mut err = crate::value::RuntimeError::new(msg);
-                    err.exception = Some(Box::new(err_val));
-                    return Some(err);
-                }
-            }
-        }
-        None
-    }
-
-    /// Compile method bodies for a given class using the bytecode compiler.
-    pub(crate) fn compile_class_methods(&mut self, class_name: &str) {
-        if let Some(class_def) = self.registry_mut().classes.get_mut(class_name) {
-            Self::compile_methods_for_map(&mut class_def.methods, class_name);
-        }
-    }
-
-    /// Compile method bodies for a given role.
-    pub(crate) fn compile_role_methods(&mut self, role_name: &str) {
-        if let Some(role_def) = self.registry_mut().roles.get_mut(role_name) {
-            Self::compile_methods_for_map(&mut role_def.methods, role_name);
-        }
-    }
-
-    pub(crate) fn smart_match_values(&mut self, left: &Value, right: &Value) -> bool {
-        self.smart_match(left, right)
-    }
-
-    pub(crate) fn eval_sequence_values(
-        &mut self,
-        left: Value,
-        right: Value,
-        exclude_end: bool,
-    ) -> Result<Value, RuntimeError> {
-        let result = self.eval_sequence(left, right, exclude_end)?;
-        // The `...` operator returns a Seq in Raku, not a List/Array.
-        // Convert finite Array results to Seq; LazyList results stay as-is.
-        match result {
-            Value::Array(items, _) => Ok(Value::Seq(std::sync::Arc::new(items.to_vec()))),
-            other => Ok(other),
-        }
-    }
-
-    pub(crate) fn resolve_code_var(&self, name: &str) -> Value {
-        let normalized_name = Self::normalize_categorical_operator_name(name);
-        if (normalized_name.starts_with("infix:<")
-            || normalized_name.starts_with("prefix:<")
-            || normalized_name.starts_with("postfix:<"))
-            && normalized_name.ends_with('>')
-        {
-            return Value::Routine {
-                package: Symbol::intern("GLOBAL"),
-                name: Symbol::intern(&normalized_name),
-                is_regex: false,
-            };
-        }
-        // Handle package-qualified names: strip pseudo-package prefixes and
-        // resolve the bare function name.
-        let bare_name = Self::strip_pseudo_packages(name);
-        let has_packages = bare_name != name;
-        let lookup_name = bare_name.strip_prefix('*').unwrap_or(bare_name);
-        if bare_name == "?ROUTINE" {
-            // Skip pointy-block entries to find the enclosing routine
-            let entry = self
-                .routine_stack
-                .iter()
-                .rev()
-                .find(|frame| frame.name != "<pointy-block>");
-            if let Some(frame) = entry {
-                // Anonymous subs are pushed with "<anon>" as the sentinel name.
-                // Return the block_stack Sub directly so callers can invoke it.
-                if frame.name.is_empty() || frame.name == "<anon>" {
-                    if let Some(val) = self.block_stack.last().cloned()
-                        && matches!(val, Value::Sub(_))
-                    {
-                        return val;
-                    }
-                    return Value::Nil;
-                }
-                return Value::Routine {
-                    package: Symbol::intern(&frame.package),
-                    name: Symbol::intern(&frame.name),
-                    is_regex: false,
-                };
-            }
-            return Value::Nil;
-        }
-        // When SETTING:: (or similar) pseudo-packages are present, resolve to
-        // the builtin directly — these refer to the outer setting scope, not
-        // user-defined overrides.
-        // When pseudo-package qualifiers are present (e.g. SETTING::), resolve
-        // to the builtin directly, bypassing user-defined overrides.
-        if has_packages && Self::is_builtin_function(lookup_name) {
-            return Value::Routine {
-                package: Symbol::intern("GLOBAL"),
-                name: Symbol::intern(lookup_name),
-                is_regex: false,
-            };
-        }
-        // For &-sigil private attribute access (e.g. &!m), the attribute
-        // value is stored in env as "!m" (not "&!m"), so check directly.
-        if bare_name.starts_with('!')
-            && let Some(val) = self.env.get(bare_name)
-        {
-            return val.clone();
-        }
-        // Check if stored as a variable first (my &f = ...)
-        let var_key = format!("&{}", bare_name);
-        if let Some(val) = self.env.get(&var_key) {
-            // Upgrade WeakSub references (e.g., &?BLOCK) to strong Sub
-            if let Value::WeakSub(weak) = val {
-                return match weak.upgrade() {
-                    Some(strong) => Value::Sub(strong),
-                    None => Value::Nil,
-                };
-            }
-            return val.clone();
-        }
-        // `return` is a control-flow keyword that also resolves as &return
-        // so that it can be rebound (proxied return pattern).
-        if bare_name == "return" {
-            return Value::Routine {
-                package: Symbol::intern("GLOBAL"),
-                name: Symbol::intern("return"),
-                is_regex: false,
-            };
-        }
-        // Look up as a function reference (including multi subs).
-        // When pseudo-packages are present (e.g. OUR::, GLOBAL::), also check
-        // our_scoped_functions for subs defined in EVAL that don't leak into
-        // the regular functions map.
-        let def = self
-            .resolve_function(lookup_name)
-            .map(|a| (*a).clone())
-            .or_else(|| {
-                if has_packages {
-                    let fq = format!("{}::{}", self.current_package(), lookup_name);
-                    self.registry()
-                        .our_scoped_functions
-                        .get(&Symbol::intern(&fq))
-                        .map(|d| (**d).clone())
-                        .or_else(|| {
-                            let global_fq = format!("GLOBAL::{}", lookup_name);
-                            self.registry()
-                                .our_scoped_functions
-                                .get(&Symbol::intern(&global_fq))
-                                .map(|d| (**d).clone())
-                        })
-                } else {
-                    None
-                }
-            });
-        let is_multi = if def.is_none() && !self.has_proto(lookup_name) {
-            // Check if there are multi-dispatch variants (stored with arity/type suffixes)
-            let prefix_local = format!("{}::{}/", self.current_package(), lookup_name);
-            let prefix_global = format!("GLOBAL::{}/", lookup_name);
-            self.registry().functions.keys().any(|k| {
-                let ks = k.resolve();
-                ks.starts_with(&prefix_local) || ks.starts_with(&prefix_global)
-            })
-        } else {
-            false
-        };
-        if is_multi {
-            // Multi subs: create a Sub that captures all candidates so the
-            // callable works even after the defining scope exits.
-            let candidates = self.resolve_all_multi_candidates(lookup_name);
-            let mut candidate_subs = Vec::new();
-            for cand in &candidates {
-                let captured_env = self.env.clone();
-                let sub_val = Value::make_sub(
-                    cand.package,
-                    cand.name,
-                    cand.params.clone(),
-                    cand.param_defs.clone(),
-                    cand.body.clone(),
-                    cand.is_rw,
-                    captured_env,
-                );
-                candidate_subs.push(sub_val);
-            }
-            let mut dispatcher_env = self.env.clone();
-            dispatcher_env.insert(
-                "__mutsu_multi_dispatch_candidates".to_string(),
-                Value::Array(
-                    std::sync::Arc::new(crate::value::ArrayData::new(candidate_subs)),
-                    crate::value::ArrayKind::List,
-                ),
-            );
-            dispatcher_env.insert(
-                "__mutsu_multi_dispatch_name".to_string(),
-                Value::str(lookup_name.to_string()),
-            );
-            Value::make_sub(
-                Symbol::intern(&self.current_package()),
-                Symbol::intern(lookup_name),
-                Vec::new(),
-                Vec::new(),
-                Vec::new(),
-                false,
-                dispatcher_env,
-            )
-        } else if self.has_proto(lookup_name)
-            || self.resolve_token_defs(lookup_name).is_some()
-            || self.has_proto_token(lookup_name)
-        {
-            Value::Routine {
-                package: Symbol::intern(&self.current_package()),
-                name: Symbol::intern(lookup_name),
-                is_regex: self.resolve_token_defs(lookup_name).is_some()
-                    || self.has_proto_token(lookup_name),
-            }
-        } else if let Some(def) = def {
-            let mut captured_env = self.env.clone();
-            if let Some(ref return_type) = def.return_type {
-                captured_env.insert(
-                    "__mutsu_return_type".to_string(),
-                    Value::str(return_type.clone()),
-                );
-            }
-            if def.is_method {
-                captured_env.insert(
-                    "__mutsu_callable_type".to_string(),
-                    Value::str_from("Method"),
-                );
-            }
-            let empty_sig = def.empty_sig;
-            let mut sub_val = Value::make_sub(
-                def.package,
-                def.name,
-                def.params,
-                def.param_defs,
-                def.body,
-                def.is_rw,
-                captured_env,
-            );
-            // Preserve empty_sig from the FunctionDef so that arity checks
-            // (e.g. sort rejecting 0-arity callables) work correctly.
-            if empty_sig && let Value::Sub(ref data) = sub_val {
-                let mut new_data = (**data).clone();
-                new_data.empty_sig = true;
-                sub_val = Value::Sub(Arc::new(new_data));
-            }
-            sub_val
-        } else if Self::is_builtin_function(lookup_name) {
-            Value::Routine {
-                package: Symbol::intern("GLOBAL"),
-                name: Symbol::intern(lookup_name),
-                is_regex: false,
-            }
-        } else if bare_name.starts_with('*') {
-            // Dynamic code vars (&*foo) can point to routines that are resolved
-            // at call time (including builtins not listed in is_builtin_function).
-            Value::Routine {
-                package: Symbol::intern("GLOBAL"),
-                name: Symbol::intern(lookup_name),
-                is_regex: false,
-            }
-        } else {
-            Value::Nil
-        }
-    }
-
-    /// Strip pseudo-package prefixes (SETTING::, OUTER::, CALLER::, CORE::, etc.)
-    /// from a qualified name and return the final bare function name.
-    pub(crate) fn strip_pseudo_packages(name: &str) -> &str {
-        let pseudo = [
-            "SETTING", "CALLER", "OUTER", "CORE", "GLOBAL", "MY", "OUR", "DYNAMIC", "UNIT",
-        ];
-        let mut rest = name;
-        loop {
-            let mut found = false;
-            for pkg in &pseudo {
-                if let Some(after) = rest.strip_prefix(pkg)
-                    && let Some(after) = after.strip_prefix("::")
-                {
-                    rest = after;
-                    found = true;
-                    break;
-                }
-            }
-            if !found {
-                break;
-            }
-        }
-        rest
-    }
-
-    pub(crate) fn routine_stack_top(&self) -> Option<&super::RoutineFrame> {
-        self.routine_stack.last()
-    }
-
-    pub(crate) fn routine_stack(&self) -> &[super::RoutineFrame] {
-        &self.routine_stack
-    }
-
-    /// Push a new routine frame. `line` and `file` record the call-site
-    /// in the *caller* (the line/file where this function was called from).
-    pub(crate) fn push_routine_with_location(
-        &mut self,
-        package: String,
-        name: String,
-        line: Option<u32>,
-        file: Option<String>,
-    ) {
-        self.routine_stack.push(super::RoutineFrame {
-            package,
-            name,
-            line,
-            file,
-            is_method: false,
-            is_block: false,
-        });
-    }
-
-    pub(crate) fn push_method_routine_with_location(
-        &mut self,
-        package: String,
-        name: String,
-        line: Option<u32>,
-        file: Option<String>,
-    ) {
-        self.routine_stack.push(super::RoutineFrame {
-            package,
-            name,
-            line,
-            file,
-            is_method: true,
-            is_block: false,
-        });
-    }
-
-    /// Push a block/closure routine frame.
-    pub(crate) fn push_block_routine_with_location(
-        &mut self,
-        package: String,
-        name: String,
-        line: Option<u32>,
-        file: Option<String>,
-    ) {
-        self.routine_stack.push(super::RoutineFrame {
-            package,
-            name,
-            line,
-            file,
-            is_method: false,
-            is_block: true,
-        });
-    }
-
-    pub(crate) fn pop_routine(&mut self) {
-        self.routine_stack.pop();
-    }
-
-    pub(crate) fn block_stack_top(&self) -> Option<&Value> {
-        self.block_stack.last()
-    }
-
-    pub(crate) fn push_block(&mut self, val: Value) {
-        self.block_stack.push(val);
-    }
-
-    pub(crate) fn pop_block(&mut self) {
-        self.block_stack.pop();
-    }
-
-    /// Stringify a value, calling the `.Str` method for Instance and Package types.
-    pub(crate) fn stringify_value(&mut self, value: Value) -> Result<String, RuntimeError> {
-        match &value {
-            Value::Instance { .. } | Value::Package(_) => {
-                let result = self.call_method_with_values(value, "Str", vec![])?;
-                Ok(result.to_string_value())
-            }
-            _ => Ok(value.to_string_value()),
-        }
-    }
-
-    /// Check if a value can respond to a given method name.
-    pub(crate) fn value_can_method(&mut self, value: &Value, method: &str) -> bool {
-        // Check builtin 0-arg method (covers most built-in methods)
-        if crate::builtins::native_method_0arg(value, crate::symbol::Symbol::intern(method))
-            .is_some()
-        {
-            return true;
-        }
-        // For instances, check class methods
-        if let Value::Instance { class_name, .. } = value
-            && self.class_has_method(&class_name.resolve(), method)
-        {
-            return true;
-        }
-        // Universal methods available on all values
-        matches!(
-            method,
-            "WHAT"
-                | "say"
-                | "print"
-                | "put"
-                | "gist"
-                | "Str"
-                | "Int"
-                | "Num"
-                | "Bool"
-                | "Numeric"
-                | "Real"
-                | "so"
-                | "not"
-                | "defined"
-                | "isa"
-                | "can"
-                | "does"
-                | "ACCEPTS"
-                | "raku"
-                | "perl"
-                | "clone"
-                | "new"
-        )
-    }
-
-    pub(crate) fn take_value(&mut self, val: Value) -> Result<(), RuntimeError> {
-        if let Some(items) = self.gather_items.last_mut() {
-            items.push(val);
-            if let Some(Some(limit)) = self.gather_take_limits.last()
-                && items.len() >= *limit
-            {
-                return Err(RuntimeError::new(
-                    "__mutsu_lazy_gather_take_limit_reached__",
-                ));
-            }
-        }
-        Ok(())
-    }
-
-    pub(crate) fn gather_items_len(&self) -> usize {
-        self.gather_items.len()
-    }
-
-    pub(crate) fn push_gather_items(&mut self, items: Vec<Value>) {
-        self.gather_items.push(items);
-    }
-
-    pub(crate) fn pop_gather_items(&mut self) -> Option<Vec<Value>> {
-        self.gather_items.pop()
-    }
-
-    pub(crate) fn push_gather_take_limit(&mut self, limit: Option<usize>) {
-        self.gather_take_limits.push(limit);
-    }
-
-    pub(crate) fn pop_gather_take_limit(&mut self) {
-        self.gather_take_limits.pop();
-    }
-
-    /// The package currently in scope, read out of the shared `Arc<RwLock>`
-    /// handle as an owned `String`. Returns owned (not `&str`) because the value
-    /// lives behind a lock guard that must not escape the call — the guard is
-    /// dropped before returning, so no lock is held across the caller's work
-    /// (re-entry safe, mirroring the registry accessors).
-    pub(crate) fn current_package(&self) -> String {
-        self.current_package.read().unwrap().clone()
-    }
-
-    pub(crate) fn set_current_package(&mut self, pkg: String) {
-        *self.current_package.write().unwrap() = pkg;
-    }
-
-    /// Interior-mutable variant for the `&self` regex matcher: the package is
-    /// stored behind a RwLock, so a temporary switch (e.g. into a cross-package
-    /// grammar subrule's defining package while parsing its body) does not need
-    /// `&mut self`.
-    pub(crate) fn set_current_package_shared(&self, pkg: String) {
-        *self.current_package.write().unwrap() = pkg;
-    }
-
-    fn stash_symbol_key_from_env_tail(rest: &str) -> String {
-        if rest.starts_with('$')
-            || rest.starts_with('@')
-            || rest.starts_with('%')
-            || rest.starts_with('&')
-        {
-            return rest.to_string();
-        }
-        if rest.contains("::") || rest.chars().next().is_some_and(|c| c.is_ascii_uppercase()) {
-            return rest.to_string();
-        }
-        format!("${rest}")
-    }
-
-    fn stash_member_tail<'a>(key: &'a str, package: &str) -> Option<&'a str> {
-        let package = package.trim_end_matches("::");
-        if package == "GLOBAL" {
-            return Some(key);
-        }
-        let direct = format!("{package}::");
-        if let Some(rest) = key.strip_prefix(&direct) {
-            return Some(rest);
-        }
-        let needle = format!("::{package}::");
-        if let Some(idx) = key.rfind(&needle) {
-            let start = idx + needle.len();
-            return Some(&key[start..]);
-        }
-        None
-    }
-
-    fn make_stash_instance(package: &str, symbols: HashMap<String, Value>) -> Value {
-        let mut attrs = HashMap::new();
-        attrs.insert("name".to_string(), Value::str(package.to_string()));
-        attrs.insert("symbols".to_string(), Value::hash(symbols));
-        Value::make_instance(Symbol::intern("Stash"), attrs)
-    }
-
-    fn package_export_tag_parts(package: &str) -> Option<(&str, &str)> {
-        let (module, rest) = package.split_once("::EXPORT::")?;
-        if module.is_empty() || rest.is_empty() || rest.contains("::") {
-            return None;
-        }
-        Some((module, rest))
-    }
-
-    fn package_export_module(package: &str) -> Option<&str> {
-        package.strip_suffix("::EXPORT")
-    }
-
-    fn qualify_stash_name(package: &str, symbol: &str) -> String {
-        let package = package.trim_end_matches("::");
-        if package.is_empty() || package == "GLOBAL" {
-            symbol.to_string()
-        } else {
-            format!("{package}::{symbol}")
-        }
-    }
-
-    fn normalize_stash_package(package: &str) -> String {
-        let trimmed = package.trim_end_matches("::");
-        if let Some(inner) = trimmed
-            .strip_prefix("GLOBAL[")
-            .and_then(|s| s.strip_suffix(']'))
-        {
-            inner.to_string()
-        } else {
-            trimmed.to_string()
-        }
-    }
-
-    fn has_package_members(&self, package: &str) -> bool {
-        let prefix = format!("{package}::");
-        self.env.keys().any(|k| k.starts_with(&prefix))
-            || self
-                .registry()
-                .functions
-                .keys()
-                .any(|k| k.resolve().starts_with(&prefix))
-            || self
-                .registry()
-                .classes
-                .keys()
-                .any(|k| k.starts_with(&prefix))
-            || self.exported_subs.contains_key(package)
-            || self.exported_vars.contains_key(package)
-    }
-
-    fn stash_lookup_symbol(stash: &Value, key: &str) -> Option<Value> {
-        let Value::Instance {
-            class_name,
-            attributes,
-            ..
-        } = stash
-        else {
-            return None;
-        };
-        if class_name != "Stash" {
-            return None;
-        }
-        let map = attributes.as_map();
-        let Value::Hash(symbols) = map.get("symbols")? else {
-            return None;
-        };
-        if let Some(value) = symbols.get(key) {
-            return Some(value.clone());
-        }
-        if !key.starts_with('$')
-            && !key.starts_with('@')
-            && !key.starts_with('%')
-            && !key.starts_with('&')
-        {
-            let scalar = format!("${key}");
-            if let Some(value) = symbols.get(&scalar) {
-                return Some(value.clone());
-            }
-        }
-        None
-    }
-
-    /// Check if a name is a known pseudo-package (MY, OUTER, CORE, etc.)
-    pub(crate) fn is_pseudo_package_name(name: &str) -> bool {
-        matches!(
-            name,
-            "SETTING"
-                | "CALLER"
-                | "CALLERS"
-                | "OUTER"
-                | "OUTERS"
-                | "CORE"
-                | "GLOBAL"
-                | "MY"
-                | "OUR"
-                | "DYNAMIC"
-                | "UNIT"
-                | "LEXICAL"
-                | "CLIENT"
-        )
-    }
-
-    fn no_such_symbol_failure(name: &str) -> Value {
-        let mut ex_attrs = HashMap::new();
-        ex_attrs.insert(
-            "message".to_string(),
-            Value::str(format!("No such symbol '{name}'")),
-        );
-        // A failed symbolic lookup (`::('NoSuchName')`, `::('')`) is
-        // X::NoSuchSymbol in Raku, not a bare X::AdHoc. `symbol` carries the name.
-        ex_attrs.insert("symbol".to_string(), Value::str(name.to_string()));
-        let exception = Value::make_instance(Symbol::intern("X::NoSuchSymbol"), ex_attrs);
-        let mut failure_attrs = HashMap::new();
-        failure_attrs.insert("exception".to_string(), exception);
-        failure_attrs.insert("handled".to_string(), Value::Bool(false));
-        Value::make_instance(Symbol::intern("Failure"), failure_attrs)
-    }
-
-    pub(crate) fn resolve_indirect_type_name(&self, name: &str) -> Value {
-        if name.is_empty() {
-            return Self::no_such_symbol_failure(name);
-        }
-        // Symbols loaded via `$*REPO.need(...)` stay invisible to `::('Name')`
-        // until they are merged into GLOBAL with `merge-symbols`.
-        if self.cur_repo.pending_global_symbols.contains(name) {
-            return Self::no_such_symbol_failure(name);
-        }
-        // Pseudo-package names like MY, CORE, OUTER, CALLER, etc. should
-        // resolve to Package values so that .WHO can produce the stash.
-        if Self::is_pseudo_package_name(name) {
-            return Value::Package(Symbol::intern(name));
-        }
-        if let Some(code_name) = name.strip_prefix('&') {
-            let val = self.resolve_code_var(code_name);
-            // When the code variable is not found via ::('&name'), return a
-            // Failure (like Raku's X::NoSuchSymbol) so that attempting to use
-            // the result throws an exception.
-            if matches!(val, Value::Nil) {
-                return Self::no_such_symbol_failure(name);
-            }
-            return val;
-        }
-        // Scalars are stored without the `$` sigil in the env; strip it for lookup.
-        if let Some(bare) = name.strip_prefix('$')
-            && let Some(value) = self.env.get(bare)
-            && !matches!(value, Value::Nil)
-        {
-            return value.clone();
-        }
-        if let Some(value) = self.env.get(name)
-            && !matches!(value, Value::Nil)
-            // Skip `my`-scoped package items for indirect type lookup (::())
-            // since they should not be visible outside their declaring scope.
-            && !self.is_my_scoped_package_item(name)
-        {
-            return value.clone();
-        }
-        // Fallback: check persistent `our`-scoped variables (constants, `our` decls)
-        // which may have been removed from the lexical env by block-scope restoration.
-        if let Some(bare) = name.strip_prefix('$')
-            && let Some(value) = self.our_vars.get(bare)
-            && !matches!(value, Value::Nil)
-        {
-            return value.clone();
-        }
-        if let Some(value) = self.our_vars.get(name)
-            && !matches!(value, Value::Nil)
-        {
-            return value.clone();
-        }
-        // Look up well-known numerical constants
-        match name {
-            "e" | "\u{1D452}" => return Value::Num(std::f64::consts::E),
-            "pi" => return Value::Num(std::f64::consts::PI),
-            "tau" | "\u{03C4}" => return Value::Num(std::f64::consts::TAU),
-            _ => {}
-        }
-        if !self.method_class_stack.is_empty() && self.loaded_modules.contains(name) {
-            return Value::Package(Symbol::intern(name));
-        }
-
-        // Check if the full compound name (e.g. "IO::Path") is a known type
-        // before splitting on "::".
-        if name.contains("::")
-            && (crate::runtime::utils::is_known_compound_type(name)
-                || self.has_class(name)
-                || self.is_role(name))
-        {
-            return Value::Package(Symbol::intern(name));
-        }
-
-        let mut parts = name.split("::").filter(|part| !part.is_empty());
-        let Some(first) = parts.next() else {
-            return Self::no_such_symbol_failure(name);
-        };
-
-        let mut current = if let Some(value) = self.env.get(first)
-            && !matches!(value, Value::Nil)
-        {
-            value.clone()
-        } else if crate::runtime::utils::is_known_type_constraint(first)
-            || (name.contains("::")
-                && (self.has_package_members(first)
-                    || self.has_class(first)
-                    || self.is_role(first)))
-        {
-            Value::Package(Symbol::intern(first))
-        } else {
-            return Self::no_such_symbol_failure(name);
-        };
-
-        for part in parts {
-            current = match &current {
-                Value::Package(package) => {
-                    let stash = self.package_stash_value(&package.resolve());
-                    if let Some(value) = Self::stash_lookup_symbol(&stash, part) {
-                        value
-                    } else {
-                        return Self::no_such_symbol_failure(name);
-                    }
-                }
-                Value::Instance { class_name, .. } if class_name == "Stash" => {
-                    if let Some(value) = Self::stash_lookup_symbol(&current, part) {
-                        value
-                    } else {
-                        return Self::no_such_symbol_failure(name);
-                    }
-                }
-                _ => return Self::no_such_symbol_failure(name),
-            };
-        }
-
-        current
-    }
-
-    pub(crate) fn package_stash_value(&self, package: &str) -> Value {
-        let package_name = Self::normalize_stash_package(package);
-
-        // PROCESS:: pseudo-package: exposes process-level dynamic variables
-        // like $*PROGRAM, $*PID, %*ENV, @*ARGS, etc.
-        // PROCESS::<$PROGRAM> looks up key "$PROGRAM" in the stash.
-        if package_name == "PROCESS" {
-            let mut symbols: HashMap<String, Value> = HashMap::new();
-            for (key, val) in self.env.iter() {
-                let key_s = key.resolve();
-                // Dynamic variables stored as "*NAME" map to "$NAME" in the stash
-                if let Some(name) = key_s.strip_prefix('*')
-                    && !name.contains("::")
-                {
-                    symbols.insert(format!("${name}"), val.clone());
-                }
-                // Hash dynamic variables stored as "%*NAME" map to "%NAME" in the stash
-                if let Some(name) = key_s.strip_prefix("%*")
-                    && !name.contains("::")
-                {
-                    symbols.insert(format!("%{name}"), val.clone());
-                }
-                // Array dynamic variables stored as "@*NAME" map to "@NAME" in the stash
-                if let Some(name) = key_s.strip_prefix("@*")
-                    && !name.contains("::")
-                {
-                    symbols.insert(format!("@{name}"), val.clone());
-                }
-            }
-            return Self::make_stash_instance(package, symbols);
-        }
-
-        if let Some((module, tag)) = Self::package_export_tag_parts(package) {
-            let mut symbols: HashMap<String, Value> = HashMap::new();
-            if let Some(subs) = self.exported_subs.get(module) {
-                for (name, tags) in subs {
-                    if tag != "ALL" && !tags.contains(tag) {
-                        continue;
-                    }
-                    let fq = format!("{module}::{name}");
-                    symbols.insert(format!("&{name}"), self.resolve_code_var(&fq));
-                }
-            }
-            if let Some(vars) = self.exported_vars.get(module) {
-                for (name, tags) in vars {
-                    if tag != "ALL" && !tags.contains(tag) {
-                        continue;
-                    }
-                    let fq = format!("{module}::{name}");
-                    let val = self
-                        .env
-                        .get(&fq)
-                        .cloned()
-                        .or_else(|| self.env.get(name).cloned())
-                        .unwrap_or(Value::Nil);
-                    symbols.insert(name.clone(), val);
-                }
-            }
-            return Self::make_stash_instance(package, symbols);
-        }
-
-        if let Some(module) = Self::package_export_module(&package_name) {
-            let mut tags = std::collections::BTreeSet::new();
-            if let Some(subs) = self.exported_subs.get(module) {
-                for tagset in subs.values() {
-                    tags.extend(tagset.iter().cloned());
-                }
-            }
-            if let Some(vars) = self.exported_vars.get(module) {
-                for tagset in vars.values() {
-                    tags.extend(tagset.iter().cloned());
-                }
-            }
-            let mut symbols: HashMap<String, Value> = HashMap::new();
-            for tag in tags {
-                symbols.insert(
-                    tag.clone(),
-                    Value::Package(Symbol::intern(&Self::qualify_stash_name(
-                        &package_name,
-                        &tag,
-                    ))),
-                );
-            }
-            return Self::make_stash_instance(package, symbols);
-        }
-
-        let mut symbols: HashMap<String, Value> = HashMap::new();
-
-        for (key, val) in self.env.iter() {
-            let key_s = key.resolve();
-            if key_s.starts_with("__mutsu_callable_id::") {
-                continue;
-            }
-            if (package_name == "MY" || package_name == "GLOBAL")
-                && self.should_hide_from_my_global_stash(&key_s)
-            {
-                continue;
-            }
-            // Skip env entries hidden from package stash lookups (transitive deps)
-            if package_name != "MY"
-                && package_name != "GLOBAL"
-                && self.package_stash_hidden.contains(&key_s)
-            {
-                continue;
-            }
-            // Skip my-scoped items (they should not appear in the package stash)
-            if self.is_my_scoped_package_item(&key_s) {
-                continue;
-            }
-            if let Some(rest) = Self::stash_member_tail(&key_s, &package_name) {
-                let stash_key = Self::stash_symbol_key_from_env_tail(rest);
-                symbols.insert(stash_key, val.clone());
-            }
-        }
-
-        for (key, def) in &self.registry().functions {
-            let key_s = key.resolve();
-            let Some(rest) = Self::stash_member_tail(&key_s, &package_name) else {
-                continue;
-            };
-            let base = rest.split('/').next().unwrap_or(rest);
-            if base.is_empty() || base.contains("::") || base.contains(':') {
-                continue;
-            }
-            // Skip my-scoped subs (they should not appear in the package stash)
-            let fq_base = format!("{}::{}", package_name, base);
-            if self.is_my_scoped_package_item(&fq_base) {
-                continue;
-            }
-            symbols
-                .entry(format!("&{base}"))
-                .or_insert_with(|| Value::Routine {
-                    package: def.package,
-                    name: def.name,
-                    is_regex: false,
-                });
-        }
-
-        for class_name in self.registry().classes.keys() {
-            let class_short = class_name
-                .rsplit_once("::")
-                .map(|(_, short)| short)
-                .unwrap_or(class_name.as_str());
-            if (package_name == "MY" || package_name == "GLOBAL")
-                && (self.need_hidden_classes.contains(class_name)
-                    || self.need_hidden_classes.contains(class_short))
-            {
-                continue;
-            }
-            // Skip classes hidden from package stash lookups (transitive deps)
-            if package_name != "MY"
-                && package_name != "GLOBAL"
-                && self.package_stash_hidden.contains(class_name)
-            {
-                continue;
-            }
-            // Skip my-scoped classes (they should not appear in the package stash)
-            if self.is_my_scoped_package_item(class_name) {
-                continue;
-            }
-            let Some(rest) = Self::stash_member_tail(class_name, &package_name) else {
-                continue;
-            };
-            if rest.is_empty() {
-                continue;
-            }
-            if let Some((head, _)) = rest.split_once("::") {
-                symbols.entry(head.to_string()).or_insert_with(|| {
-                    Value::Package(Symbol::intern(&Self::qualify_stash_name(
-                        &package_name,
-                        head,
-                    )))
-                });
-                continue;
-            }
-            symbols.entry(rest.to_string()).or_insert_with(|| {
-                Value::Package(Symbol::intern(&Self::qualify_stash_name(
-                    &package_name,
-                    rest,
-                )))
-            });
-        }
-        for role_name in self.registry().roles.keys() {
-            // Skip roles hidden from package stash lookups (transitive deps)
-            if package_name != "MY"
-                && package_name != "GLOBAL"
-                && self.package_stash_hidden.contains(role_name)
-            {
-                continue;
-            }
-            let Some(rest) = Self::stash_member_tail(role_name, &package_name) else {
-                continue;
-            };
-            if rest.is_empty() {
-                continue;
-            }
-            if let Some((head, _)) = rest.split_once("::") {
-                symbols.entry(head.to_string()).or_insert_with(|| {
-                    Value::Package(Symbol::intern(&Self::qualify_stash_name(
-                        &package_name,
-                        head,
-                    )))
-                });
-                continue;
-            }
-            symbols.entry(rest.to_string()).or_insert_with(|| {
-                Value::Package(Symbol::intern(&Self::qualify_stash_name(
-                    &package_name,
-                    rest,
-                )))
-            });
-        }
-
-        if self.exported_subs.contains_key(&package_name)
-            || self.exported_vars.contains_key(&package_name)
-        {
-            symbols.entry("EXPORT".to_string()).or_insert_with(|| {
-                Value::Package(Symbol::intern(&Self::qualify_stash_name(
-                    &package_name,
-                    "EXPORT",
-                )))
-            });
-        }
-
-        Self::make_stash_instance(package, symbols)
-    }
-
-    pub(crate) fn push_end_phaser(&mut self, body: Vec<Stmt>) {
-        let captured_env = self.env.clone();
-        self.end_phasers.push((body, captured_env));
-    }
-
-    /// Return the number of currently registered END phasers.
-    pub(crate) fn end_phaser_count(&self) -> usize {
-        self.end_phasers.len()
-    }
-
-    /// Update captured envs for END phasers registered since `start_idx`
-    /// with the current values from the given env.  This ensures that END
-    /// phasers see final variable values rather than stale copies captured
-    /// at registration time (e.g., inside closures or block scopes).
-    pub(crate) fn update_end_phaser_envs(&mut self, start_idx: usize, current_env: &Env) {
-        for phaser in self.end_phasers[start_idx..].iter_mut() {
-            let captured = &mut phaser.1;
-            for (k, v) in current_env {
-                if captured.contains_key_sym(*k) {
-                    captured.insert_sym(*k, v.clone());
-                }
-            }
-        }
-    }
-
-    /// Update captured envs of ALL END phasers, but only for the specified
-    /// variable names.  Used after closure calls to propagate changes to
-    /// captured variables without overwriting unrelated variables.
-    pub(crate) fn update_end_phaser_envs_for_keys(
-        &mut self,
-        keys: &std::collections::HashSet<&str>,
-        current_env: &Env,
-    ) {
-        for phaser in self.end_phasers.iter_mut() {
-            let captured = &mut phaser.1;
-            for k in keys {
-                if captured.contains_key(k)
-                    && let Some(v) = current_env.get(k)
-                {
-                    captured.insert(k.to_string(), v.clone());
-                }
-            }
-        }
-    }
-
-    /// Register an END phaser site_id. Returns true if this is the first
-    /// registration (phaser should be pushed), false if already registered.
-    pub(crate) fn register_end_phaser_site(&mut self, site_id: u64) -> bool {
-        self.end_phaser_sites.insert(site_id)
-    }
-
-    pub(crate) fn snapshot_routine_registry(&self) -> RoutineRegistrySnapshot {
-        // Single guard for all six reads (avoids stacking read guards).
-        let registry = self.registry();
-        (
-            registry.functions.clone(),
-            registry.proto_functions.clone(),
-            registry.token_defs.clone(),
-            registry.proto_subs.clone(),
-            registry.proto_tokens.clone(),
-            registry.our_scoped_functions.keys().copied().collect(),
-        )
-    }
-
-    pub(crate) fn restore_routine_registry(&mut self, snapshot: RoutineRegistrySnapshot) {
-        self.restore_routine_registry_impl(snapshot, false);
-    }
-
-    pub(crate) fn restore_routine_registry_eval(&mut self, snapshot: RoutineRegistrySnapshot) {
-        self.restore_routine_registry_impl(snapshot, true);
-    }
-
-    fn restore_routine_registry_impl(&mut self, snapshot: RoutineRegistrySnapshot, is_eval: bool) {
-        let (functions, proto_functions, token_defs, proto_subs, proto_tokens, our_scoped_keys) =
-            snapshot;
-        // Collect our-scoped functions that were newly added during this block
-        // (not present in the snapshot) that need to persist after scope restoration.
-        // Preserve functions defined in the current package (original behavior)
-        // OR functions newly registered during this block (so nested package
-        // declarations like `{ package Foo { our sub bar {} } }` survive). We
-        // distinguish "newly registered during this block" by tracking the set
-        // of our_scoped_functions keys that existed at snapshot time.
-        let current_pkg = self.current_package();
-        let mut new_our: Vec<(Symbol, std::sync::Arc<FunctionDef>)> = Vec::new();
-        // Collect under a read guard, which drops before the writes below
-        // (read->write on the same lock would deadlock).
-        {
-            let registry = self.registry();
-            for (key, def) in &registry.our_scoped_functions {
-                if functions.contains_key(key) {
-                    continue;
-                }
-                let def_pkg = def.package.resolve();
-                // Always preserve same-package subs (original behavior).
-                if def_pkg == current_pkg {
-                    new_our.push((*key, def.clone()));
-                    continue;
-                }
-                // Also preserve subs that were newly added to our_scoped_functions
-                // during this block (i.e. their key was not in the snapshot). This
-                // covers nested `package Foo { our sub bar {} }` blocks. Subs from
-                // module loading (`use Foo`) are typically already in the snapshot
-                // by the time the block is restored, so they are not preserved here.
-                if !our_scoped_keys.contains(key) {
-                    new_our.push((*key, def.clone()));
-                }
-            }
-        }
-        let mut registry = self.registry_mut();
-        registry.functions = functions;
-        registry.proto_functions = proto_functions;
-        registry.token_defs = token_defs;
-        registry.proto_subs = proto_subs;
-        registry.proto_tokens = proto_tokens;
-        // Re-apply only newly added our-scoped functions so they survive block scope exit.
-        // In EVAL context, our-scoped functions should NOT leak into the outer lexical
-        // scope — they remain accessible only via OUR:: pseudo-package resolution.
-        if !is_eval {
-            for (key, def) in new_our {
-                registry.functions.insert(key, def);
-            }
-        }
-        drop(registry);
-        // The routine registry just changed: a lexical (`my sub`) registered
-        // inside the block was removed. Invalidate the name-keyed resolution
-        // caches so a subsequent call to that name re-resolves against the
-        // restored registry rather than returning the now-out-of-scope
-        // CompiledFunction the first in-block call cached (the compiled-function
-        // map is program-global and still contains the lexical sub's body).
-        self.fn_resolve_gen += 1;
-        self.method_resolve_cache.clear();
-        self.last_method_resolve = None;
-        self.fast_method_cache.clear();
-        self.multi_resolve_cache.clear();
-        self.multi_type_cacheable.clear();
-    }
-
-    pub(crate) fn push_block_scope_depth(&mut self) {
-        self.block_scope_depth += 1;
-    }
-
-    pub(crate) fn pop_block_scope_depth(&mut self) {
-        self.block_scope_depth = self.block_scope_depth.saturating_sub(1);
-    }
-
-    /// Push a saved variable value for `let`/`temp` scope management.
-    /// `is_temp`: true for `temp` (always restore), false for `let` (restore on failure only).
-    pub(crate) fn let_saves_push(&mut self, name: String, value: Value, is_temp: bool) {
-        // Take an independent snapshot: an instance value must be deep-copied so
-        // a later in-place mutation through its shared cell does not corrupt the
-        // saved-for-restore value (Phase 3, Stage 1).
-        self.let_saves
-            .push((name, value.into_temp_snapshot(), is_temp));
-    }
-
-    /// Restore one `let`/`temp` save. For an instance, write the saved
-    /// attributes back into the *live* shared cell of the variable's current
-    /// binding so every alias sees the restoration and object identity (id +
-    /// cell) is preserved, rather than rebinding the name to a detached copy.
-    fn restore_let_value(&mut self, name: String, restored: Value) {
-        // A boxed (shared-cell) binding: write the restored value THROUGH the live
-        // cell so the owner's local slot (which holds the same Arc) sees the
-        // restoration, rather than replacing the env entry with a detached plain
-        // value (which would strand the slot's stale cell). The `let`-save already
-        // captured the inner value (see `exec_let_save_op`). Covers named-sub
-        // captured-outer boxing (docs/captured-outer-cell-sharing.md). Gated on the
-        // same toggle as the boxing it supports, so the default build is
-        // byte-identical to before.
-        if let Some(Value::ContainerRef(arc)) = self.env.get(&name).cloned() {
-            arc.lock().unwrap().clone_from(&restored);
-            return;
-        }
-        // Non-cell restore writes the saved value into `env` by name only; record
-        // it so the block-exit op refreshes the owning local slot precisely
-        // (`apply_pending_rw_writeback`) instead of relying on the blanket
-        // env→locals pull (env_dirty-removal substrate). The owner is usually this
-        // same frame (a bare `{ temp $x = … }` block), but a `temp`/`let` inside a
-        // nested callee owns the slot a frame up — record on both the drop-on-miss
-        // and retain-on-miss lists so either reaches it.
-        self.pending_rw_writeback_sources.push(name.clone());
-        self.record_caller_var_writeback(&name);
-        if let Value::Instance {
-            attributes: saved_attrs,
-            ..
-        } = &restored
-        {
-            // The current binding for `name` holds the live instance (the same
-            // shared cell that was mutated during the dynamic scope). Write the
-            // saved snapshot straight back into that cell; the env binding already
-            // aliases it, so no re-insert is needed.
-            if let Some(Value::Instance {
-                attributes: live_attrs,
-                ..
-            }) = self.env.get(&name)
-            {
-                live_attrs.commit_attrs(saved_attrs.to_map());
-                return;
-            }
-            self.env.insert(name, restored);
-        } else {
-            self.env.insert(name, restored);
-        }
-    }
-
-    /// Current length of let_saves stack (used as a mark).
-    pub(crate) fn let_saves_len(&self) -> usize {
-        self.let_saves.len()
-    }
-
-    /// Resolve the value to restore, applying `is default(...)` when restoring Nil.
-    fn resolve_restore_value(&self, name: &str, val: &Value) -> Value {
-        if matches!(val, Value::Nil)
-            && let Some(default) = self.var_defaults.get(name)
-        {
-            return default.clone();
-        }
-        val.clone()
-    }
-
-    /// Restore all variables from let_saves starting at `mark`, then truncate.
-    pub(crate) fn restore_let_saves(&mut self, mark: usize) {
-        for i in (mark..self.let_saves.len()).rev() {
-            let (name, old_val, _is_temp) = self.let_saves[i].clone();
-            let restored = self.resolve_restore_value(&name, &old_val);
-            self.restore_let_value(name, restored);
-        }
-        self.let_saves.truncate(mark);
-    }
-
-    /// On successful block exit: restore `temp` saves, discard `let` saves.
-    /// For `let`, only restore if the block returned an unsuccessful value.
-    pub(crate) fn resolve_let_saves_on_success(&mut self, mark: usize, success: bool) {
-        // Collect restore actions first to avoid borrow conflicts.
-        let restores: Vec<(String, Value)> = (mark..self.let_saves.len())
-            .rev()
-            .filter_map(|i| {
-                let (ref name, ref old_val, is_temp) = self.let_saves[i];
-                if is_temp || !success {
-                    Some((name.clone(), old_val.clone()))
-                } else {
-                    None
-                }
-            })
-            .collect();
-        for (name, old_val) in restores {
-            let restored = self.resolve_restore_value(&name, &old_val);
-            self.restore_let_value(name, restored);
-        }
-        self.let_saves.truncate(mark);
-    }
-
-    /// Discard let_saves from `mark` without restoring (block succeeded).
-    pub(crate) fn discard_let_saves(&mut self, mark: usize) {
-        self.let_saves.truncate(mark);
     }
 }
